normalize json filter paths across db dialects
